Refactor: FogOfWar to module
